Load lifestyle context editors on demand
