+++ tests/test-correctness-phase2.js
@@ -140,6 +140,7 @@ const pwaAppShellAssets = [
   '/js/context-card-editor-ui.js',
   '/js/context-card-medical-history-editor.js',
   '/js/context-card-lifestyle-editors.js',
+  '/js/context-card-lifestyle-editors-impl.js',
   '/js/supplement-impact.js',
   '/js/commit-hash.js',
   '/js/focus-card.js',
