+++ tests/test-prelab.js
@@ -142,7 +142,7 @@ const onboardingRuntimeSrc = read('js/onboarding-view-runtime.js');
   const ccSrc = read('js/context-cards.js');
   const healthDotsSrc = read('js/context-card-health-dots.js');
   const ctxEditorSrc = read('js/context-card-editor-ui.js');
-  const lifestyleEditorSrc = read('js/context-card-lifestyle-editors.js');
+  const lifestyleEditorSrc = read('js/context-card-lifestyle-editors-impl.js');
   const medicalHistoryEditorSrc = read('js/context-card-medical-history-editor.js');
 
   assert('context-cards imports getActiveData', ccSrc.includes("import { saveImportedData, getActiveData } from './data.js'"),
