+++ tests/test-sync.js
@@ -140,7 +140,7 @@ await import('../js/settings.js');
   const lightEnvSrc = await fetchWithRetry('js/light-env.js');
   const supplementsSrc = await fetchWithRetry('js/supplements.js');
   const notesSrc = await fetchWithRetry('js/notes.js');
-  const contextCardLifestyleEditorsSrc = await fetchWithRetry('js/context-card-lifestyle-editors.js');
+  const contextCardLifestyleEditorsSrc = await fetchWithRetry('js/context-card-lifestyle-editors-impl.js');
   const chatSummariesSrc = await fetchWithRetry('js/chat-summaries.js');
   const markerDetailModalSrc = [
     await fetchWithRetry('js/marker-detail-modal.js'),
