+++ tsconfig.checkjs.json
@@ -100,6 +100,7 @@
     "js/context-card-health-dots.js",
     "js/context-card-lifestyle-runtime.js",
     "js/context-card-lifestyle-editors.js",
+    "js/context-card-lifestyle-editors-impl.js",
     "js/context-card-medical-history-editor.js",
     "js/context-card-summaries.js",
     "js/context-source-registry.js",
